Eliminate log post links in Atom XSLT page
Having the Atom feed page link to each article is confusing because the
distinction between it and the actual log index page becomes fuzzy.

The Atom feed page only exists to preview what the feed includes, not to
act as its own index of posts.

This change pulls the focus away from the posts and focuses the user on
the instructions to copy the URL into their feed reader.